refactor: prepare data model for packet of protocol_1 controls
